+++ application/controllers/Info.php
@@ -23,31 +23,52 @@ class Info extends CI_Controller {
      *   
      */
     public function index() {
-        // $this->data['pagebody'] = 'welcome';
-        // $this->render(); 
         header("Content-type: application/json");
         echo json_encode("{scenario: medieval}");
-        // echo json_encode($this->Categories->all());
     }
 
     /**
-     *
+     * Displays the categories in JSON format.
+     * Displays all entries if none specified.
      */
-    public function category() {
-        $this->Categories->all();
+    public function category($key = null) {
+        $data = $this->Categories->all();
+        if ($key != null) {
+            echo json_encode($data[$key]);
+        } else {
+            foreach ($data as $row) {
+                echo json_encode($row);
+            }
+        }
     }
 
     /**
-     *
+     * Displays the accessories in JSON format.
+     * Displays all entries if none specified.
      */
-    public function catalog() {
-        $this->Accessories->all();
+    public function catalog($key = null) {
+        $data = $this->Accessories->all();
+        if ($key != null) {
+            echo json_encode($data[$key]);
+        } else {
+            foreach ($data as $row) {
+                echo json_encode($row);
+            }
+        }
     }
 
     /**
-     *
+     * Displays the equipment sets in JSON format.
+     * Displays all entries if none specified.
      */
-    public function bundle() {
-        $this->Equipment_sets->all();
+    public function bundle($key = null) {
+        $data = $this->Equipment_sets->all();
+        if ($key != null) {
+            echo json_encode($data[$key]);
+        } else {
+            foreach ($data as $row) {
+                echo json_encode($row);
+            }
+        }
     }
 }